modules/IdTranslator/requirements.txt
azure-iot-device
black; python_version >= '3.6'
pre-commit
flake8
paho.mqtt>=1.5.1
requests
requests-unixsocket
typing-extensions; python_version <= '3.8'
watchdog[watchmedo]